+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,13 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
-
     <application
         android:allowBackup="true"
         android:icon="@mipmap/crap_logo"
         android:logo="@mipmap/crap_logo"
         android:label="CrapApp"
         android:theme="@style/AppTheme"
-        android:name=".App"
-        >
+        android:name=".App">
         <activity
             android:name=".MainActivity"
             android:label="CrapApp"
