+++ app/src/main/java/com/example/android/mobilecourse/App.java
@@ -1,6 +1,4 @@
 package com.example.android.mobilecourse;
-
-
 import android.app.Application;
 
 import com.google.firebase.auth.FirebaseAuth;
@@ -13,6 +11,7 @@ public class App extends Application {
     private MovieApi movieService;
     private FirebaseAuth auth;
 
+    @Override
     public void onCreate()
     {
         super.onCreate();
