+++ app/src/main/java/com/example/android/mobilecourse/CustomAdapter.java
@@ -1,20 +1,13 @@
 package com.example.android.mobilecourse;
-
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-
 import com.squareup.picasso.Picasso;
-
 import java.util.List;
-
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class CustomAdapter extends RecyclerView.Adapter<CustomAdapter.CustomViewHolder> {
+public class CustomAdapter extends RecyclerView.Adapter<CustomViewHolder> {
 
     private List<Movie> movieList;
 
@@ -24,15 +17,13 @@ public class CustomAdapter extends RecyclerView.Adapter<CustomAdapter.CustomView
 
     @NonNull
     @Override
-    public CustomAdapter.CustomViewHolder onCreateViewHolder(@NonNull final ViewGroup parent, final int viewType) {
-
+    public CustomViewHolder onCreateViewHolder(@NonNull final ViewGroup parent, final int viewType) {
         View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.custom_view, parent, false);
         return new CustomViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(final CustomViewHolder holder, final int position) {
-        holder.textTitle.setText(movieList.get(position).getTitle());
         holder.textDesctiption.setText(movieList.get(position).getDescription());
         Picasso.get().load(movieList.get(position).getPoster()).into(holder.poster);
         holder.movieYear.setText(String.format("Year: %s", Integer.toString(movieList.get(position).getYear())));
@@ -43,20 +34,4 @@ public class CustomAdapter extends RecyclerView.Adapter<CustomAdapter.CustomView
         return movieList.size();
     }
 
-     class CustomViewHolder extends RecyclerView.ViewHolder {
-
-        private TextView textTitle;
-        private TextView textDesctiption;
-        private ImageView poster;
-        private TextView movieYear;
-
-        private CustomViewHolder(final View itemView) {
-            super(itemView);
-
-            textDesctiption = itemView.findViewById(R.id.custom_desctription);
-            poster = itemView.findViewById(R.id.custom_imageView);
-            textTitle = itemView.findViewById(R.id.custom_title);
-            movieYear = itemView.findViewById(R.id.custom_year);
-        }
-    }
 }
