+++ app/src/main/java/com/example/android/mobilecourse/Movie.java
@@ -1,6 +1,5 @@
 package com.example.android.mobilecourse;
 
-
 public class Movie {
 
     private String title;
