+++ app/src/main/java/com/example/android/mobilecourse/MovieApi.java
@@ -1,12 +1,10 @@
 package com.example.android.mobilecourse;
 
-
 import java.util.List;
 import retrofit2.Call;
 import retrofit2.http.GET;
 
 public interface MovieApi {
-
     @GET("/custom_movies/")
     Call<List<Movie>> getAllMovies();
 }
