+++ app/src/main/res/layout/activity_main.xml
@@ -6,10 +6,7 @@
     android:gravity="center_horizontal"
     android:orientation="vertical"
     tools:context=".MainActivity"
-    android:paddingTop="24dp"
-    android:paddingLeft="24dp"
-    android:paddingRight="24dp"
-    android:paddingBottom="24dp"
+    android:padding = "24dp"
     android:id="@+id/linearLayout">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
@@ -26,13 +23,6 @@
 
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
-    <Button
-        android:id="@+id/wellcome_sign_out_button"
-        android:layout_width="200dp"
-        android:layout_marginTop="20dp"
-        android:layout_height="wrap_content"
-        android:textColor="@color/white"
-        android:text="Sign out"
-        tools:ignore="HardcodedText" />
+
 
 </LinearLayout>
